Next bit of NSInteger cleanup for -gui.
Fix all of the current warnings in Gorm so that it can be used to test
NSInteger things.



git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/trunk@36039 72102866-910b-0410-8b05-ffd578937521

+++ ColorPickers/GSNamedColorPicker.m
@@ -228,18 +228,18 @@ createRowsForColumn: (NSInteger)column
 {
 }
 
-- (int) numberOfItemsInComboBox: (NSComboBox *)aComboBox
+- (NSInteger) numberOfItemsInComboBox: (NSComboBox *)aComboBox
 {
    return [lists count];
 }
 
 - (id) comboBox: (NSComboBox *)aComboBox 
-objectValueForItemAtIndex: (int)index
+objectValueForItemAtIndex: (NSInteger)index
 {
    return [(NSColorList*)[lists objectAtIndex: index] name];
 }
 
-- (unsigned int) comboBox: (NSComboBox *)aComboBox
+- (NSUInteger) comboBox: (NSComboBox *)aComboBox
 indexOfItemWithStringValue: (NSString *)string
 {
    return [lists indexOfObject: [NSColorList colorListNamed: string]];
