+++ Headers/AppKit/NSComboBox.h
@@ -48,8 +48,8 @@
 - (float) itemHeight;
 - (void) setItemHeight: (float)itemHeight;
 
-- (int) numberOfVisibleItems;
-- (void) setNumberOfVisibleItems: (int)visibleItems;
+- (NSInteger) numberOfVisibleItems;
+- (void) setNumberOfVisibleItems: (NSInteger)visibleItems;
 
 - (void) reloadData;
 - (void) noteNumberOfItemsChanged;
@@ -57,13 +57,13 @@
 - (BOOL) usesDataSource;
 - (void) setUsesDataSource: (BOOL)flag;
 
-- (void) scrollItemAtIndexToTop: (int)index;
-- (void) scrollItemAtIndexToVisible: (int)index;
+- (void) scrollItemAtIndexToTop: (NSInteger)index;
+- (void) scrollItemAtIndexToVisible: (NSInteger)index;
 
-- (void) selectItemAtIndex: (int)index;
-- (void) deselectItemAtIndex: (int)index;
-- (int) indexOfSelectedItem;
-- (int) numberOfItems;
+- (void) selectItemAtIndex: (NSInteger)index;
+- (void) deselectItemAtIndex: (NSInteger)index;
+- (NSInteger) indexOfSelectedItem;
+- (NSInteger) numberOfItems;
 
 /* These two methods can only be used when usesDataSource is YES */
 - (id) dataSource;
@@ -72,14 +72,14 @@
 /* These methods can only be used when usesDataSource is NO */
 - (void) addItemWithObjectValue: (id)object;
 - (void) addItemsWithObjectValues: (NSArray *)objects;
-- (void) insertItemWithObjectValue: (id)object atIndex:(int)index;
+- (void) insertItemWithObjectValue: (id)object atIndex:(NSInteger)index;
 - (void) removeItemWithObjectValue: (id)object;
-- (void) removeItemAtIndex: (int)index;
+- (void) removeItemAtIndex: (NSInteger)index;
 - (void) removeAllItems;
 - (void) selectItemWithObjectValue: (id)object;
-- (id) itemObjectValueAtIndex: (int)index;
+- (id) itemObjectValueAtIndex: (NSInteger)index;
 - (id) objectValueOfSelectedItem;
-- (int) indexOfItemWithObjectValue: (id)object;
+- (NSInteger) indexOfItemWithObjectValue: (id)object;
 - (NSArray *) objectValues;
 
 #if OS_API_VERSION(GS_API_MACOSX, GS_API_LATEST)
@@ -95,9 +95,9 @@
 @end
 
 @interface NSObject (NSComboBoxDataSource)
-- (int) numberOfItemsInComboBox: (NSComboBox *)aComboBox;
-- (id) comboBox: (NSComboBox *)aComboBox objectValueForItemAtIndex:(int)index;
-- (unsigned int) comboBox: (NSComboBox *)aComboBox 
+- (NSInteger) numberOfItemsInComboBox: (NSComboBox *)aComboBox;
+- (id) comboBox: (NSComboBox *)aComboBox objectValueForItemAtIndex:(NSInteger)index;
+- (NSUInteger) comboBox: (NSComboBox *)aComboBox 
   indexOfItemWithStringValue: (NSString *)string;
 #if OS_API_VERSION(GS_API_MACOSX, GS_API_LATEST)
 /* text completion */
