+++ Headers/AppKit/NSForm.h
@@ -45,15 +45,15 @@
 //
 - (NSFormCell*)addEntry:(NSString*)title;
 - (NSFormCell*)insertEntry:(NSString*)title
-		    atIndex:(int)index;
-- (void)removeEntryAtIndex:(int)index;
+		    atIndex:(NSInteger)index;
+- (void)removeEntryAtIndex:(NSInteger)index;
 - (void)setInterlineSpacing:(float)spacing;
 
 //
 // Finding Indices
 //
-- (int)indexOfCellWithTag:(int)aTag;
-- (int)indexOfSelectedItem;
+- (NSInteger)indexOfCellWithTag:(NSInteger)aTag;
+- (NSInteger)indexOfSelectedItem;
 
 //
 // Modifying Graphic Attributes 
@@ -72,17 +72,17 @@
 //
 // Getting a Cell 
 //
-- (id)cellAtIndex:(int)index;
+- (id)cellAtIndex:(NSInteger)index;
 
 //
 // Displaying a Cell
 //
-- (void)drawCellAtIndex:(int)index;
+- (void)drawCellAtIndex:(NSInteger)index;
 
 //
 // Editing Text 
 //
-- (void)selectTextAtIndex:(int)index;
+- (void)selectTextAtIndex:(NSInteger)index;
 
 //
 // Resizing the Form
